updated root background, navbar responsiveness, fixed speaker card height, analytics cards ui
