+++ src/components/HeroSection.jsx
@@ -63,10 +63,10 @@ const HeroSection = () => {
       <div>
         <div className="mt-8">
           {/* todo, work in progress */}
-          <div className="text-2xl font-sans font-medium ">
+          <div className="text-lg sm:text-xl text-center font-sans font-medium ">
             Summer School: 29th July to 3rd August
           </div>
-          <div className="text-2xl font-sans font-medium ">
+          <div className="text-lg sm:text-xl text-center font-sans font-medium ">
             Hackathon: 7th August to 9th August
           </div>
         </div>
@@ -96,19 +96,25 @@ const HeroSection = () => {
               </span>
             </button>
           </a>
-          <div
+
+          {/* <div
             className="apply-button"
             data-hackathon-slug="web3ssh"
             data-button-theme="light"
             style={{ height: "44px", width: "312px" }}
-          ></div>
-          {/*<a href='https://web3ssh.devfolio.co/' target='_blank'>
-            <button className='w-full h-full bg-[#3770FF] flex items-center justify-center p-4 text-md rounded-md'>
-              <img src='https://avatars.githubusercontent.com/u/38809367?v=4' height="30px" width="30px" />
-              Apply for Hackathon (via Devfolio)
+            bg-[#3770FF]
+          ></div> */}
+
+          <a href='https://web3ssh.devfolio.co/' target='_blank'>
+            <button className='w-full h-full bg-[#3770FF] flex items-center justify-center p-3 text-lg rounded-full'>
+              <span>
+              <img src='https://avatars.githubusercontent.com/u/38809367?v=4' height="30px" width="30px"  className='inline'/>
+              Apply with Devfolio
+              </span>
+              <span className='text-lg'>&nbsp; &lt;Hackathon&gt; </span>
             </button>
           </a>
-          */}
+
         </div>
       </div>
     </div>
