+++ src/components/Speakers.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
-import SpeakerCard from './daywiseoverview/SpeakerCardV2';
+import SpeakerCard from './daywiseoverview/SpeakerCard';
 import speakers from '../data/speakers.json';
 
 export default function SpeakersPanel() {
   return (
-    <div className="w-full mb-[120px] px-4 sm:px-6 lg:px-8 flex flex-col justify-center items-center">
+    <section id="speakers" className="w-full mb-[120px] px-4 sm:px-6 lg:px-8 flex flex-col justify-center items-center">
       <div
-        id="speakers"
         className="text-white text-4xl md:text-6xl mb-10 lg:mb-20 font-sans font-medium text-center"
       >
         <h1>Speakers</h1>
@@ -22,6 +21,6 @@ export default function SpeakersPanel() {
           />
         ))}
       </div>
-    </div>
+    </section>
   );
 }
